Fix warnings C4267 conversion from 'size_t' to 'int', possible loss of data

+++ fmatvec/var_vector.h
@@ -298,7 +298,7 @@ namespace fmatvec {
     }
 
   template <class AT>
-    inline Vector<Var,AT>::Vector(const std::vector<AT> &v) : Matrix<General,Var,Fixed<1>,AT>(v.size(),1) {
+    inline Vector<Var,AT>::Vector(const std::vector<AT> &v) : Matrix<General,Var,Fixed<1>,AT>(static_cast<int>(v.size()),1) {
       if(size()>0) memcpy(&operator()(0), &v[0], sizeof(AT)*size());
     }
